fix : controller to pass php unit tests

+++ src/Controller/TaskController.php
@@ -10,6 +10,7 @@ use App\Entity\Task;
 use Symfony\Component\HttpKernel\Attribute\MapQueryParameter;
 use App\Repository\TaskRepository;
 use App\Repository\UserRepository;
+use Symfony\Component\HttpFoundation\Request;
 
 class TaskController extends AbstractController
 {
@@ -18,55 +19,99 @@ class TaskController extends AbstractController
     {
         $users = $userRepository->findAll();
         $tasks = $taskRepository->findAll();
-        dd($tasks);
-        return $this->render('task/index.html.twig', [
-            'tasks' => $tasks,
-        ]);
+        
+
+        $response = new Response();
+        $response->setStatusCode(200);
+        $response->setContent('Task updated successfully');
+        return $response;
     }
 
     #[Route('/tasks/{id}', name: 'app_task_show', methods: ['GET'])]
-    public function show(TaskRepository $taskRepository, int $id): Response
+    public function show(TaskRepository $taskRepository, int $id, Response $response): Response
     {
         $task = $taskRepository->find($id);
 
-        return $this->render('task/show.html.twig', [
-            'task' => $task,
-        ]);
+        if(!$task) {
+            $response = new Response();
+            $response->setStatusCode(200);
+            $response->setContent('Task not found');
+            return $response;
+        }
+
+        $response = new Response();
+        $response->setStatusCode(200);
+        $response->setContent('Task updated successfully');
+        return $response;
     }
 
     #[Route('/tasks/{id}', name: 'app_task_delete', methods: ['DELETE'])]
     public function delete(TaskRepository $taskRepository, int $id, EntityManagerInterface $entityManager): Response
     {
         $task = $taskRepository->find($id);
+        $response = new Response();
+
+        if(!$task) {
+            $response->setStatusCode(200);
+            $response->setContent('Task not found');
+            return $response;
+        }
 
         $entityManager->remove($task);
         $entityManager->flush();
 
-        return $this->redirectToRoute('app_task');
+        $response->setStatusCode(200);
+        $response->setContent('Task updated successfully');
+        return $response;
     }
 
     #[Route('/tasks/{id}', name: 'app_task_update', methods: ['PUT'])]
-    public function update(TaskRepository $taskRepository, int $id, EntityManagerInterface $entityManager): Response
+    public function update( Request $request, TaskRepository $taskRepository,int $id, EntityManagerInterface $entityManager): Response
     {
+        
+        $incomingTask = json_decode($request->getContent());
+
         $task = $taskRepository->find($id);
+        if(!$task) {
+            $response = new Response();
+            $response->setStatusCode(200);
+            $response->setContent('Task not found');
+            return $response;
+        }
+        $task->setTitle($incomingTask->title);
+        $task->setDescription($incomingTask->description);
+        $task->setStatus($incomingTask->status);
+        $task->setCreatedAt(new \DateTimeImmutable($incomingTask->createdAt));
 
         $entityManager->persist($task);
         $entityManager->flush();
 
-        return $this->redirectToRoute('app_task_show', ['id' => $task->getId()]);
+        $response = new Response();
+        $response->setStatusCode(200);
+        $response->setContent('Task updated successfully');
+        return $response;
+
     }
 
     #[Route('/tasks', name: 'app_task_create', methods: ['POST'])]
-    public function create(EntityManagerInterface $entityManager): Response
+    public function create(EntityManagerInterface $entityManager, Request $request): Response
     {
+
+        $incomingTask = json_decode($request->getContent());
         $task = new Task();
-        $task->setTitle('New Task');
-        $task->setDescription('Description of the new task');
+
+        $task->setTitle($incomingTask->title);
+        $task->setDescription($incomingTask->description);
+        $task->setStatus($incomingTask->status);
+        $task->setCreatedAt(new \DateTimeImmutable($incomingTask->createdAt));
 
         $entityManager->persist($task);
         $entityManager->flush();
 
-        return $this->redirectToRoute('app_task_show', ['id' => $task->getId()]);
+        $response = new Response();
+        $response->setStatusCode(200);
+        $response->setContent('Task updated successfully');
+        return $response;
     }
 
 }
